refactor: 파일 구조 수정 및 페이지네이션 함수 커스텀
